# docs/swagger.yaml
definitions:
...
  login.LoginUserPassRequest:
    properties:
      name:
        type: string
      password:
        maxLength: 16
        minLength: 8
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  response.ApiResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /v1/file/delete/{filename}:
    delete:
      consumes:
      - application/json
      operationId: DeleteFile
      parameters:
      - description: File Name
        in: path
        name: filename
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Uploaded File
      tags:
      - file
  /v1/file/show/{filename}:
    get:
      consumes:
      - application/json
      operationId: ShowFile
      parameters:
      - description: File Name
        in: path
        name: filename
        required: true
        type: string
      produces:
      - '*/*'
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Show Uploaded File
      tags:
      - file
  /v1/file/upload:
    post:
      consumes:
      - '*/*'
      operationId: UploadFile
      parameters:
      - description: File Upload Request
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Upload File
      tags:
      - file
...
  /v1/user-login/login:
    post:
      consumes:
      - application/json
      operationId: Login
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/login.LoginUserPassRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      summary: Login
      tags:
      - usermanagement
swagger: "2.0"